+++ crates/jarl/tests/integration/statistics.rs
@@ -18,6 +18,8 @@ any(is.na(x))
 any(is.na(x))
 any(is.na(x))
 any(is.na(x))
+f(x = 1, x = 2)
+stop('x')
 ",
         ),
         ("test2.R", "mean(x <- 1)"),
@@ -28,6 +30,8 @@ any(is.na(x))
             .command()
             .arg("check")
             .arg(".")
+            .arg("--select")
+            .arg("any_is_na,condition_call,duplicated_arguments")
             .arg("--statistics")
             .run()
             .normalize_os_executable_name(),
@@ -37,9 +41,11 @@ any(is.na(x))
     exit_code: 1
     ----- stdout -----
        12 [*] any_is_na
-        1 [ ] implicit_assignment
+        1 [ ] duplicated_arguments
+        1 [^] condition_call
 
-    Rules with `[*]` have an automatic fix.
+    Rules with `[*]` have an automatic safe fix.
+    Rules with `[^]` have an automatic unsafe fix.
 
     ----- stderr -----
     "